+++ js/app.js
@@ -1,9 +1,9 @@
 // app.js
 // アプリケーションの初期化のみ担当
 
-import { showPastedText } from './showPaste.js';
-import { initCopyButton } from './copyButtonHandler.js';
-import { initDisplayArea } from './displayAreaHandler.js';
+import { showPastedText } from './views/showPaste.js';
+import { initCopyButton } from './controllers/copyButtonHandler.js';
+import { initDisplayArea } from './controllers/displayAreaHandler.js';
 
 window.addEventListener('DOMContentLoaded', () => {
     const confirmBtn = document.getElementById('confirm-btn');
