+++ js/infrastructure/clipboard.js
